begin cleaning up and separating
