+++ src/jecs/src/Shared/CommonJecsHooks/changed.lua
@@ -8,7 +8,7 @@
 local require = require(script.Parent.Parent.loader).load(script)
 
 local ImmediateTypes = require("ImmediateTypes")
-local getOrCreateHookState = require("ImmediateHookUtils").getOrCreateHookState
+local getOrCreateHookState = require("JecsImmediateHookUtils").getOrCreateHookState
 
 return function(rt: ImmediateTypes.ImmediateRuntime)
 	return function(value: any, dis: any?, runFirst: boolean?, onlyOnEqualTo: any?)
