+++ src/jecs/src/Shared/CommonJecsHooks/conditionSustained.lua
@@ -9,7 +9,7 @@
 local require = require(script.Parent.Parent.loader).load(script)
 
 local ImmediateTypes = require("ImmediateTypes")
-local getOrCreateHookState = require("ImmediateHookUtils").getOrCreateHookState
+local getOrCreateHookState = require("JecsImmediateHookUtils").getOrCreateHookState
 
 return function(rt: ImmediateTypes.ImmediateRuntime)
 	return function(condition: boolean, dis: any?)
