+++ src/jecs/src/Shared/CommonJecsHooks/noise.lua
@@ -20,7 +20,7 @@
 local require = require(script.Parent.Parent.loader).load(script)
 
 local ImmediateTypes = require("ImmediateTypes")
-local getOrCreateHookState = require("ImmediateHookUtils").getOrCreateHookState
+local getOrCreateHookState = require("JecsImmediateHookUtils").getOrCreateHookState
 
 return function(rt: ImmediateTypes.ImmediateRuntime)
 	return function(dis: any?, min: number?, max: number?, frequency: number?): number
