+++ src/jecs/src/Shared/CommonJecsHooks/random.lua
@@ -6,7 +6,7 @@
 local require = require(script.Parent.Parent.loader).load(script)
 
 local ImmediateTypes = require("ImmediateTypes")
-local getOrCreateHookState = require("ImmediateHookUtils").getOrCreateHookState
+local getOrCreateHookState = require("JecsImmediateHookUtils").getOrCreateHookState
 
 return function(rt: ImmediateTypes.ImmediateRuntime)
 	return function(min: number?, max: number?, dis: any?, seed: number?)
