+++ src/jecs/src/Shared/CommonJecsHooks/throttle.lua
@@ -8,7 +8,7 @@
 local require = require(script.Parent.Parent.loader).load(script)
 
 local ImmediateTypes = require("ImmediateTypes")
-local getOrCreateHookState = require("ImmediateHookUtils").getOrCreateHookState
+local getOrCreateHookState = require("JecsImmediateHookUtils").getOrCreateHookState
 
 return function(rt: ImmediateTypes.ImmediateRuntime)
 	return function(seconds: number, dis: any?, delayOnFirstCall: boolean?)
