+++ src/jecs/src/Shared/CommonJecsHooks/useTieInterface.lua
@@ -9,7 +9,7 @@ local require = require(script.Parent.Parent.loader).load(script)
 
 local ImmediateTypes = require("ImmediateTypes")
 local TieRealms = require("TieRealms")
-local getOrCreateHookState = require("ImmediateHookUtils").getOrCreateHookState
+local getOrCreateHookState = require("JecsImmediateHookUtils").getOrCreateHookState
 
 return function(rt: ImmediateTypes.ImmediateRuntime)
 	return function(dis, instance: Instance, tieInterfaceName: string, realm: TieRealms.TieRealm?)
