+++ src/jecs/src/Shared/CommonJecsHooks/value.lua
@@ -10,7 +10,7 @@ local require = require(script.Parent.Parent.loader).load(script)
 
 local ImmediateTypes = require("ImmediateTypes")
 local ValueObject = require("ValueObject")
-local getOrCreateHookState = require("ImmediateHookUtils").getOrCreateHookState
+local getOrCreateHookState = require("JecsImmediateHookUtils").getOrCreateHookState
 
 return function(rt: ImmediateTypes.ImmediateRuntime)
 	return function(initialValue: any, dis: any?): ValueObject.ValueObject<any>
