+++ src/jecs/src/Shared/Immediate/JecsImmediateHooksComponents.lua
@@ -5,17 +5,10 @@
 
 local require = require(script.Parent.loader).load(script)
 
-local Jecs = require("Jecs")
 local Jecst = require("Jecst")
-local Maid = require("Maid")
 
 return function(world: Jecst.World)
 	local newComponents = {
-		Name = Jecs.Name,
-		Previous = Jecs.Rest,
-		ChildOf = Jecs.ChildOf,
-
-		Maid = world:component() :: Jecst.Id<Maid.Maid>,
 		MetaHookState = world:component() :: Jecst.Id<{
 			filename: string,
 			line: number,
@@ -26,18 +19,6 @@ return function(world: Jecst.World)
 		}>,
 		HookState = world:component() :: Jecst.Id<{ [any]: any }>,
 		HookRuntimeBuffer = world:component() :: Jecst.Id<nil>,
-		_ChangeCallbacks = world:component() :: Jecst.Id<{
-			_counter: number,
-			callbacks: { [string]: (entity: Jecst.Entity, id: number, data: any) -> () },
-		}>,
-		_AddCallbacks = world:component() :: Jecst.Id<{
-			_counter: number,
-			callbacks: { [string]: (entity: Jecst.Entity, id: number, data: any) -> () },
-		}>,
-		_RemoveCallbacks = world:component() :: Jecst.Id<{
-			_counter: number,
-			callbacks: { [string]: (entity: Jecst.Entity, id: number, delete: boolean?) -> () },
-		}>,
 	}
 	return newComponents
 end
